Exercise generator split.
